Add scripts with formatted code
